Rename the reindex_worker to the reindex_request_creator
