+++ reindexer/Makefile
@@ -5,7 +5,7 @@ STACK_ROOT 	= reindexer
 SBT_DOCKER_LIBRARIES =
 SBT_NO_DOCKER_LIBRARIES =
 
-SBT_APPS 	= reindex_worker
+SBT_APPS 	= reindex_request_creator
 ECS_TASKS 	=
 LAMBDAS 	= reindex_shard_generator
