+++ reindexer/reindex_request_creator/src/main/resources/logback.xml
