+++ reindexer/reindex_request_creator/src/main/scala/uk/ac/wellcome/platform/reindex_request_creator/models/ReindexJob.scala
@@ -1,0 +1,11 @@
+package uk.ac.wellcome.platform.reindex_request_creator.models
+
+/** A request to identify all the shards in the table that need reindexing.
+  *
+  * @param shardId Name of the shard to reindex.
+  * @param desiredVersion Version to reindex everything in this shard to.
+  */
+case class ReindexJob(
+  shardId: String,
+  desiredVersion: Int
+)
