+++ reindexer/reindex_request_creator/src/main/scala/uk/ac/wellcome/platform/reindex_request_creator/models/ReindexRequest.scala
@@ -1,0 +1,13 @@
+package uk.ac.wellcome.platform.reindex_request_creator.models
+
+/** A request to update a single record in SourceData to [[desiredVersion]].
+  *
+  * This is sent to an SNS topic, and handled by another application.
+  *
+  * @param id ID of the record to update (e.g. "miro/A1234567")
+  * @param desiredVersion The reindex version to set on the record.
+  */
+case class ReindexRequest(
+  id: String,
+  desiredVersion: Int
+)
