+++ reindexer/reindex_request_creator/src/test/scala/uk/ac/wellcome/platform/reindex_request_creator/ReindexRequestCreatorFeatureTest.scala
@@ -1,0 +1,107 @@
+package uk.ac.wellcome.platform.reindex_request_creator
+
+import com.gu.scanamo.Scanamo
+import org.scalatest.concurrent.{Eventually, ScalaFutures}
+import org.scalatest.{FunSpec, Matchers}
+import uk.ac.wellcome.messaging.sns.NotificationMessage
+import uk.ac.wellcome.messaging.test.fixtures.SQS.Queue
+import uk.ac.wellcome.messaging.test.fixtures.{SNS, SQS}
+import uk.ac.wellcome.models.Id
+import uk.ac.wellcome.platform.reindex_request_creator.models.{
+  ReindexJob,
+  ReindexRequest
+}
+import uk.ac.wellcome.storage.test.fixtures.LocalDynamoDb.Table
+import uk.ac.wellcome.storage.test.fixtures.LocalDynamoDbVersioned
+import uk.ac.wellcome.test.utils.ExtendedPatience
+import uk.ac.wellcome.utils.JsonUtil._
+
+case class TestRecord(
+  id: String,
+  someData: String,
+  version: Int,
+  reindexShard: String,
+  reindexVersion: Int
+) extends Id
+
+class ReindexRequestCreatorFeatureTest
+    extends FunSpec
+    with Matchers
+    with Eventually
+    with ExtendedPatience
+    with fixtures.Server
+    with LocalDynamoDbVersioned
+    with SNS
+    with SQS
+    with ScalaFutures {
+
+  val currentVersion = 1
+  val desiredVersion = 5
+
+  val shardName = "shard"
+
+  private def createReindexableData(queue: Queue,
+                                    table: Table): Seq[ReindexRequest] = {
+    val numberOfRecords = 4
+
+    val testRecords = (1 to numberOfRecords).map(i => {
+      TestRecord(
+        id = s"id$i",
+        version = 1,
+        someData = "A ghastly gharial ganking a green golem.",
+        reindexShard = shardName,
+        reindexVersion = currentVersion
+      )
+    })
+
+    //TODO re-factor shared test state here into fixture method
+    testRecords.foreach(Scanamo.put(dynamoDbClient)(table.name)(_))
+
+    val reindexJob = ReindexJob(
+      shardId = shardName,
+      desiredVersion = desiredVersion
+    )
+
+    val sqsMessage = NotificationMessage(
+      Subject = "",
+      Message = toJson(reindexJob).get,
+      TopicArn = "topic",
+      MessageId = "message"
+    )
+
+    sqsClient.sendMessage(queue.url, toJson(sqsMessage).get)
+
+    testRecords.map { record =>
+      ReindexRequest(
+        id = record.id,
+        desiredVersion = desiredVersion
+      )
+    }
+  }
+
+  it("sends a notification for every record that needs a reindex") {
+    withLocalSqsQueue { queue =>
+      withLocalDynamoDbTable { table =>
+        withLocalSnsTopic { topic =>
+          val flags = snsLocalFlags(topic) ++ dynamoDbLocalEndpointFlags(table) ++ sqsLocalFlags(
+            queue)
+
+          withServer(flags) { _ =>
+            val expectedRecords =
+              createReindexableData(queue, table)
+
+            eventually {
+              val actualRecords: Seq[ReindexRequest] =
+                listMessagesReceivedFromSNS(topic)
+                  .map { _.message }
+                  .map { fromJson[ReindexRequest](_).get }
+                  .distinct
+
+              actualRecords should contain theSameElementsAs expectedRecords
+            }
+          }
+        }
+      }
+    }
+  }
+}
