+++ reindexer/reindex_shard_generator/src/.pytest_cache/v/cache/lastfailed
@@ -1,0 +1,3 @@
+{
+  "test_reindex_shard_config.py::test_create_reindex_shard[miro-A0000001-miro/001]": true
+}
