--- reindexer/reindex_worker/Dockerfile
@@ -1,11 +1,0 @@
-FROM wellcome/finatra_service_base
-
-ADD target/universal/stage /opt/docker
-
-USER root
-RUN chown -R daemon:daemon /opt/docker
-
-ENV project=reindex_worker
-
-USER daemon
-CMD ["/run.sh"]
