--- reindexer/reindex_worker/src/main/scala/uk/ac/wellcome/platform/reindex_worker/models/ReindexJob.scala
@@ -1,11 +1,0 @@
-package uk.ac.wellcome.platform.reindex_worker.models
-
-/** A request to identify all the shards in the table that need reindexing.
-  *
-  * @param shardId Name of the shard to reindex.
-  * @param desiredVersion Version to reindex everything in this shard to.
-  */
-case class ReindexJob(
-  shardId: String,
-  desiredVersion: Int
-)
