--- reindexer/reindex_worker/src/test/scala/uk/ac/wellcome/platform/reindex_worker/fixtures/Server.scala
@@ -1,25 +1,0 @@
-package uk.ac.wellcome.platform.reindex_worker.fixtures
-
-import com.twitter.finatra.http.EmbeddedHttpServer
-import org.scalatest.Suite
-import uk.ac.wellcome.monitoring.test.fixtures.CloudWatch
-import uk.ac.wellcome.platform.reindex_worker.{Server => AppServer}
-import uk.ac.wellcome.test.fixtures.TestWith
-
-trait Server extends CloudWatch { this: Suite =>
-  def withServer[R](flags: Map[String, String])(
-    testWith: TestWith[EmbeddedHttpServer, R]): R = {
-    val server: EmbeddedHttpServer = new EmbeddedHttpServer(
-      new AppServer(),
-      flags = flags ++ cloudWatchLocalFlags
-    )
-
-    server.start()
-
-    try {
-      testWith(server)
-    } finally {
-      server.close()
-    }
-  }
-}
