+++ reindexer/terraform/ecr.tf
@@ -1,4 +1,4 @@
-module "ecr_repository_reindex_worker" {
+module "ecr_repository_reindex_request_creator" {
   source = "git::https://github.com/wellcometrust/terraform.git//ecr?ref=v1.0.0"
-  name   = "reindex_worker"
+  name   = "reindex_request_creator"
 }
